terraform: facilitator -> data share processor
The facilitator is just a specifc instance of a data share processor,
and the Terraform module is not restricted to deploying facilitation
servers. This commit replaces occurences of "facilitator" with "data
share processor" throughout Terraform modules and some READMEs. The Rust
crate is still facilitator because renaming it is a little more
involved.